refactor: Made helpers internal
`IONFILEController` cannot be instantiated with custom helpers outside this library/module.
