+++ src/main/kotlin/io/ionic/libs/ionfilesystemlib/helper/IONFILEDirectoriesHelper.kt
@@ -1,9 +1,9 @@
 package io.ionic.libs.ionfilesystemlib.helper
 
-import io.ionic.libs.ionfilesystemlib.helper.internal.createDirOrFile
-import io.ionic.libs.ionfilesystemlib.helper.internal.deleteDirOrFile
-import io.ionic.libs.ionfilesystemlib.helper.internal.getMetadata
-import io.ionic.libs.ionfilesystemlib.helper.internal.prepareForCopyOrRename
+import io.ionic.libs.ionfilesystemlib.helper.common.createDirOrFile
+import io.ionic.libs.ionfilesystemlib.helper.common.deleteDirOrFile
+import io.ionic.libs.ionfilesystemlib.helper.common.getMetadata
+import io.ionic.libs.ionfilesystemlib.helper.common.prepareForCopyOrRename
 import io.ionic.libs.ionfilesystemlib.model.IONFILECreateOptions
 import io.ionic.libs.ionfilesystemlib.model.IONFILEDeleteOptions
 import io.ionic.libs.ionfilesystemlib.model.IONFILEExceptions
@@ -12,7 +12,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import java.io.File
 
-class IONFILEDirectoriesHelper {
+internal class IONFILEDirectoriesHelper {
 
     /**
      * List the contents of a directory
