+++ src/main/kotlin/io/ionic/libs/ionfilesystemlib/helper/common/IONFILECommon.kt
@@ -1,4 +1,4 @@
-package io.ionic.libs.ionfilesystemlib.helper.internal
+package io.ionic.libs.ionfilesystemlib.helper.common
 
 import android.annotation.SuppressLint
 import android.net.Uri
