+++ src/test/java/io/ionic/libs/ionfilesystemlib/common/IONFILEBaseJUnitTest.kt
@@ -3,7 +3,7 @@ package io.ionic.libs.ionfilesystemlib.common
 import android.net.Uri
 import android.os.Build
 import android.webkit.MimeTypeMap
-import io.ionic.libs.ionfilesystemlib.helper.internal.IONFILEBuildConfig
+import io.ionic.libs.ionfilesystemlib.helper.common.IONFILEBuildConfig
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.mockkObject
