+++ src/test/java/io/ionic/libs/ionfilesystemlib/helper/IONFILELocalFilesHelperTest.kt
@@ -4,7 +4,7 @@ import android.os.Build
 import app.cash.turbine.test
 import io.ionic.libs.ionfilesystemlib.common.IONFILEBaseJUnitTest
 import io.ionic.libs.ionfilesystemlib.common.LOREM_IPSUM_2800_CHARS
-import io.ionic.libs.ionfilesystemlib.helper.internal.IONFILEBuildConfig
+import io.ionic.libs.ionfilesystemlib.helper.common.IONFILEBuildConfig
 import io.ionic.libs.ionfilesystemlib.model.IONFILECreateOptions
 import io.ionic.libs.ionfilesystemlib.model.IONFILEEncoding
 import io.ionic.libs.ionfilesystemlib.model.IONFILEExceptions
